Floats for explosive impact and beacons

+++ src/Engine/Objects/Character.cpp
@@ -7054,7 +7054,7 @@ void Character::_42ECB5_CharacterAttacksActor() {
                                       target_id, &a3);
         if (character->WearsItem(ITEM_ARTIFACT_SPLITTER, ITEM_SLOT_MAIN_HAND) ||
             character->WearsItem(ITEM_ARTIFACT_SPLITTER, ITEM_SLOT_OFF_HAND))
-            _42FA66_do_explosive_impact(actor->pos.toInt() + Vec3i(0, 0, actor->height / 2), 0, 512, pParty->activeCharacterIndex());
+            _42FA66_do_explosive_impact(actor->pos + Vec3f(0, 0, actor->height / 2), 0, 512, pParty->activeCharacterIndex());
     } else if (bow_idx) {
         shooting_bow = true;
         pushSpellOrRangedAttack(SPELL_BOW_ARROW, pParty->activeCharacterIndex() - 1, CombinedSkillValue::none(), 0, 0);
@@ -7117,7 +7117,7 @@ void Character::_42ECB5_CharacterAttacksActor() {
 }
 
 //----- (0042FA66) --------------------------------------------------------
-void Character::_42FA66_do_explosive_impact(Vec3i pos, int a4, int16_t a5, int actchar) {
+void Character::_42FA66_do_explosive_impact(Vec3f pos, int a4, int16_t a5, int actchar) {
         // EXPLOSIVE IMPACT OF ARTIFACT SPLITTER
 
     // a5 is range?
@@ -7129,9 +7129,9 @@ void Character::_42FA66_do_explosive_impact(Vec3i pos, int a4, int16_t a5, int a
     a1a.spell_level = 8;
     a1a.spell_skill = CHARACTER_SKILL_MASTERY_MASTER;
     a1a.uObjectDescID = pObjectList->ObjectIDByItemID(a1a.uType);
-    a1a.vPosition = pos.toFloat();
+    a1a.vPosition = pos;
     a1a.uAttributes = 0;
-    a1a.uSectorID = pIndoor->GetSector(pos);
+    a1a.uSectorID = pIndoor->GetSector(pos.toInt());
     a1a.timeSinceCreated = 0_ticks;
     a1a.spell_target_pid = Pid();
     a1a.field_60_distance_related_prolly_lod = 0;
@@ -7473,7 +7473,7 @@ bool Character::setBeacon(int index, Duration duration) {
 
     beacon.image = render->TakeScreenshot(92, 68);
     beacon.uBeaconTime = pParty->GetPlayingTime() + duration;
-    beacon._partyPos = pParty->pos.toInt();
+    beacon._partyPos = pParty->pos;
     beacon._partyViewYaw = pParty->_viewYaw;
     beacon._partyViewPitch = pParty->_viewPitch;
     beacon.mapId = file_index;
